refactor to finish copying
